refactor: tiers DTO를 use case별로 organize하고 클래스명 개선
- tiers.dto.ts를 application/dto/get-all-tiers.dto.ts로 이동
- TierDto를 GetAllTiersResponseDto로 개명하여 use case 명확화
- 컨트롤러 import 경로를 새 위치로 업데이트
- 아키텍처 규칙에 따른 DTO 구조 개선
