+++ src/tiers/application/tiers.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 
 import { TiersService } from './tiers.service';
-import { TierDto, TiersResponseDto } from './tiers.dto';
+import { GetAllTiersResponseDto, TiersResponseDto } from './dto/get-all-tiers.dto';
 
 @ApiTags('Tiers')
 @Controller('tiers')
@@ -16,9 +16,9 @@ export class TiersController {
     description: '티어 목록 조회 성공',
     type: TiersResponseDto,
   })
-  async getAllTiers(): Promise<TierDto[]> {
+  async getAllTiers(): Promise<GetAllTiersResponseDto[]> {
     const tiers = await this.tiersService.getAllTiers();
 
-    return tiers.map((tier) => TierDto.from(tier));
+    return tiers.map((tier) => GetAllTiersResponseDto.from(tier));
   }
 }
